Final changes
Cleaned up, added some comments, removed unused methods, etc.
Gates should function fully now
Consistent theme throughout

+++ app/Http/Controllers/Admin/AdCommentController.php
@@ -10,15 +10,14 @@ use Illuminate\Support\Facades\Gate;
 class AdCommentController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Displays all comments associated with the user.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-
         $c = Comment::where('user_id', auth()->id())->first();
-        if(is_null($c)) {
+        if(is_null($c) && Gate::allows('isAdmin')) {
             $comments = Comment::where('user_id', auth()->id())->paginate(10);
             return view('admin.comments.index', ['comments' => $comments]);
         }
@@ -32,9 +31,14 @@ class AdCommentController extends Controller
         }
     }
 
+    /**
+     * Displays all comments from all users.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function indexAll()
     {
-        if(Gate::allows('loggedIn')) {
+        if(Gate::allows('isAdmin')) {
             $comments = Comment::paginate(10);
             return view('admin.comments.indexall', ['comments' => $comments]);
         } else {
@@ -42,65 +46,31 @@ class AdCommentController extends Controller
         }
     }
 
-    // /**
-    //  * Show the form for creating a new resource.
-    //  *
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function create()
-    // {
-    //     //
-    // }
-
-    // /**
-    //  * Store a newly created resource in storage.
-    //  *
-    //  * @param  \Illuminate\Http\Request  $request
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function store(Request $request)
-    // {
-    //     //
-    // }
-
-    // /**
-    //  * Display the specified resource.
-    //  *
-    //  * @param  int  $id
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function show($id)
-    // {
-    //     //
-    // }
-
     /**
-     * Show the form for editing the specified resource.
+     * Show the form for editing a comment.
      *
-     * @param  int  $id
+     * @param  Comment  $comment
      * @return \Illuminate\Http\Response
      */
     public function edit(Comment $comment)
     {
-        if(Gate::allows('ownsComment', $comment)) {
+        if(Gate::allows('ownsComment', $comment) && Gate::allows('isAdmin')) {
             return view('admin.comments.edit', ['comment' => $comment]);
-        }
-        else {
+        } else {
             return redirect()->route('home')->with('message', 'You must be the comment owner to access this page.');
         }
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update a comment.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  Comment  $comment
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Comment $comment)
     {
-        if(Gate::allows('ownsComment', $comment)) {
-
+        if(Gate::allows('ownsComment', $comment) && Gate::allows('isAdmin')) {
             Comment::find($comment->id)->update([
                 'content' => $request['content'],
                 'user_id' => auth()->id(),
@@ -109,26 +79,24 @@ class AdCommentController extends Controller
 
             session() -> flash('message', 'Comment edited.');
             return redirect() -> route('admin.comments.index');
-        }
-        else {
+        } else {
             return redirect()->route('home')->with('message', 'You must be the comment owner to access this page.');
         }
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove a comment.
      *
-     * @param  int  $id
+     * @param  Comment  $comment
      * @return \Illuminate\Http\Response
      */
     public function destroy(Comment $comment)
     {
-        if(Gate::allows('isAdmin')) {
+        if(Gate::allows('ownsComment', $comment) && Gate::allows('isAdmin')) {
             Comment::destroy($comment -> id);
             session() -> flash('message', 'Comment deleted.');
             return redirect() -> route('admin.comments.indexall');
-        }
-        else {
+        } else {
             return redirect()->route('home')->with('message', 'You must be the comment owner or an admin to delete a comment.');
         }
     }
