+++ app/Http/Controllers/User/UsCommentController.php
@@ -6,21 +6,19 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Comment;
 use Illuminate\Support\Facades\Auth;
-
 use Illuminate\Support\Facades\Gate;
 
 class UsCommentController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Displays all comments associated with the user.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-
         $c = Comment::where('user_id', auth()->id())->first();
-        if(is_null($c)) {
+        if(is_null($c) && Gate::allows('isUser')) {
             $comments = Comment::where('user_id', auth()->id())->paginate(10);
             return view('user.comments.index', ['comments' => $comments]);
         }
@@ -34,92 +32,55 @@ class UsCommentController extends Controller
         }
     }
 
-    // /**
-    //  * Show the form for creating a new resource.
-    //  *
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function create()
-    // {
-    //     //
-    // }
-
-    // /**
-    //  * Store a newly created resource in storage.
-    //  *
-    //  * @param  \Illuminate\Http\Request  $request
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function store(Request $request)
-    // {
-    //     //
-    // }
-
-    // /**
-    //  * Display the specified resource.
-    //  *
-    //  * @param  int  $id
-    //  * @return \Illuminate\Http\Response
-    //  */
-    // public function show($id)
-    // {
-    //     //
-    // }
-
     /**
-     * Show the form for editing the specified resource.
+     * Show the form for editing a comment.
      *
-     * @param  int  $id
+     * @param  Comment $comment
      * @return \Illuminate\Http\Response
      */
     public function edit(Comment $comment)
     {
-        if(Gate::allows('ownsComment', $comment)) {
+        if(Gate::allows('ownsComment', $comment) && Gate::allows('isUser')) {
             return view('user.comments.edit', ['comment' => $comment]);
-        }
-        else {
+        } else {
             return redirect()->route('home')->with('message', 'You must be the comment owner to access this page.');
         }
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update a comment.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  Comment $comment
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Comment $comment)
     {
-
-        if(Gate::allows('ownsComment', $comment)) {
+        if(Gate::allows('ownsComment', $comment) && Gate::allows('isUser')) {
             Comment::find($comment->id)->update([
                 'content' => $request['content']
             ]);
 
             session() -> flash('message', 'Comment edited.');
             return redirect() -> route('user.comments.index');
-        }
-        else {
+        } else {
             return redirect()->route('home')->with('message', 'You must be the comment owner to access this page.');
         }
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove a comment.
      *
-     * @param  int  $id
+     * @param  Comment $comment
      * @return \Illuminate\Http\Response
      */
     public function destroy(Comment $comment)
     {
-        if(Gate::allows('ownsComment', $comment)){
-        Comment::destroy($comment->id);
-        return redirect() -> route('user.comments.index');
-        }else {
+        if(Gate::allows('ownsComment', $comment) && Gate::allows('isUser')) {
+            Comment::destroy($comment->id);
+            return redirect() -> route('user.comments.index');
+        } else {
             return redirect()->route('home')->with('message', 'You must be the comment owner or an admin to delete this comment.');
-
+        }
     }
-}
-
 }
